refactor: move trigger helpers (normalizeTrigger/isBadTrigger) to engine/triggers + lift regex (B-026, B-072)
Second slice of the `services/utils.ts` demote. `normalizeTrigger`
and `isBadTrigger` are trigger-shape concerns, not service helpers —
new home is `src/engine/triggers.ts`.

Bundled with B-072: `isBadTrigger`'s two regex literals (forbidden
characters + colon-in-middle) lifted to module-level consts. The
function runs on the editor-change hot path (per separator
keystroke), so even though V8 caches regex literals, having them
named and at the top makes the intent obvious and removes one layer
between the call and the matcher.

Three import sites updated: `packages/espanso.ts`,
`services/community-packages.ts`, `ui/components/SnippetsTab.ts`.

Tests move to `src/engine/triggers.test.ts` (2 describe blocks, 9 it
blocks — kept verbatim from `services/utils.test.ts`).

No behaviour changes. tsc clean, 325/325 tests green.
